# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Revenue Management API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Frete {
  id: number;
  dataCte: string;
  origem: string;
  /** @nullable */
  transporte?: string | null;
  frota: string;
  /** @nullable */
  transp?: string | null;
  cliente: string;
  cidade: string;
  /** @nullable */
  cteNf?: string | null;
  peso: number;
  frete: number;
  pedagio: number;
  totalFrete: number;
  /** @nullable */
  dtaFrete?: string | null;
  /** @nullable */
  vencimento?: string | null;
  /** @nullable */
  obs?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface FreteInput {
  dataCte: string;
  origem: string;
  transporte?: string;
  frota: string;
  transp?: string;
  cliente: string;
  cidade: string;
  cteNf?: string;
  peso?: number;
  frete: number;
  pedagio?: number;
  dtaFrete?: string;
  vencimento?: string;
  obs?: string;
}

export interface FreteUpdate {
  dataCte?: string;
  origem?: string;
  transporte?: string;
  frota?: string;
  transp?: string;
  cliente?: string;
  cidade?: string;
  cteNf?: string;
  peso?: number;
  frete?: number;
  pedagio?: number;
  dtaFrete?: string;
  vencimento?: string;
  obs?: string;
}

export interface FreteListResponse {
  fretes: Frete[];
  total: number;
}

export interface BulkFreteInput {
  fretes: FreteInput[];
}

export interface BulkFreteResult {
  created: number;
  fretes: Frete[];
}

export interface FrotaSummary {
  frota: string;
  totalFretes: number;
  totalFrete: number;
  totalPedagio: number;
  totalGeral: number;
}

export interface Abastecimento {
  id: number;
  mes: string;
  ano: number;
  /** @nullable */
  requisicao?: string | null;
  /** @nullable */
  posto?: string | null;
  data: string;
  placa: string;
  litros: number;
  precoLitro: number;
  totalPago: number;
  /** @nullable */
  kmInicio?: number | null;
  /** @nullable */
  kmFinal?: number | null;
  /** @nullable */
  kmPercorrido?: number | null;
  /** @nullable */
  media?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface AbastecimentoInput {
  mes: string;
  ano: number;
  requisicao?: string;
  posto?: string;
  data: string;
  placa: string;
  litros: number;
  precoLitro: number;
  totalPago: number;
  kmInicio?: number;
  kmFinal?: number;
  kmPercorrido?: number;
  media?: number;
}

export interface AbastecimentoUpdate {
  mes?: string;
  ano?: number;
  requisicao?: string;
  posto?: string;
  data?: string;
  placa?: string;
  litros?: number;
  precoLitro?: number;
  totalPago?: number;
  kmInicio?: number;
  kmFinal?: number;
  kmPercorrido?: number;
  media?: number;
}

export interface AbastecimentoListResponse {
  abastecimentos: Abastecimento[];
  total: number;
}

export interface PlacaSummary {
  placa: string;
  totalAbastecimentos: number;
  totalLitros: number;
  totalPago: number;
  /** @nullable */
  mediaGeral?: number | null;
}

export interface DashboardResumo {
  totalFrete: number;
  totalPedagio: number;
  totalGeral: number;
  totalViagens: number;
  totalPeso: number;
  mediaPorViagem: number;
  /** @nullable */
  melhorFrota?: string | null;
  totalDiesel: number;
  totalLitros: number;
}

export interface FrotaRevenue {
  frota: string;
  totalFrete: number;
  totalPedagio: number;
  totalGeral: number;
  totalViagens: number;
  rank: number;
}

export interface PeriodoRevenue {
  periodo: string;
  totalFrete: number;
  totalPedagio: number;
  totalGeral: number;
  totalViagens: number;
}

export interface DieselByPlaca {
  placa: string;
  totalLitros: number;
  totalPago: number;
  /** @nullable */
  mediaGeral?: number | null;
  /** @nullable */
  kmTotal?: number | null;
  totalAbastecimentos: number;
}

export interface MensalComparativo {
  mes: string;
  frete: number;
  pedagio: number;
  diesel: number;
  lucroLiquido: number;
  viagens: number;
}

export type ListFretesParams = {
frota?: string;
cliente?: string;
origem?: string;
placa?: string;
dateFrom?: string;
dateTo?: string;
search?: string;
limit?: number;
offset?: number;
};

export type ListAbastecimentosParams = {
placa?: string;
mes?: string;
ano?: number;
search?: string;
limit?: number;
offset?: number;
};

export type GetDashboardResumoParams = {
dateFrom?: string;
dateTo?: string;
};

export type GetRevenueByFrotaParams = {
dateFrom?: string;
dateTo?: string;
};

export type GetRevenueByPeriodoParams = {
period: GetRevenueByPeriodoPeriod;
dateFrom?: string;
dateTo?: string;
};

export type GetRevenueByPeriodoPeriod = typeof GetRevenueByPeriodoPeriod[keyof typeof GetRevenueByPeriodoPeriod];


export const GetRevenueByPeriodoPeriod = {
  diario: 'diario',
  semanal: 'semanal',
  mensal: 'mensal',
  trimestral: 'trimestral',
  anual: 'anual',
} as const;

export type GetDieselByPlacaParams = {
ano?: number;
};

export type GetMensalComparativoParams = {
ano?: number;
};