# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Fleet Revenue Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /abastecimentos:
    get:
      operationId: listAbastecimentos
      tags: [abastecimentos]
      summary: List fuel records
      parameters:
        - name: placa
          in: query
          schema: { type: string }
        - name: mes
          in: query
          schema: { type: string }
        - name: ano
          in: query
          schema: { type: integer }
        - name: search
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 1000 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbastecimentoListResponse"
    post:
      operationId: createAbastecimento
      tags: [abastecimentos]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbastecimentoInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Abastecimento"
  /abastecimentos/{id}:
    put:
      operationId: updateAbastecimento
      tags: [abastecimentos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AbastecimentoUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Abastecimento"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAbastecimento
      tags: [abastecimentos]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
...
  /dashboard/resumo:
    get:
      operationId: getDashboardResumo
      tags: [dashboard]
      parameters:
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResumo"
  /dashboard/por-frota:
    get:
      operationId: getRevenueByFrota
      tags: [dashboard]
      parameters:
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FrotaRevenue"
  /dashboard/por-periodo:
    get:
      operationId: getRevenueByPeriodo
      tags: [dashboard]
      parameters:
        - name: period
          in: query
          required: true
          schema:
            type: string
            enum: [diario, semanal, mensal, trimestral, anual]
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeriodoRevenue"
  /dashboard/diesel-por-placa:
    get:
      operationId: getDieselByPlaca
      tags: [dashboard]
      parameters:
        - name: ano
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DieselByPlaca"
  /dashboard/mensal:
    get:
      operationId: getMensalComparativo
      tags: [dashboard]
      parameters:
        - name: ano
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MensalComparativo"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Abastecimento:
      type: object
      properties:
        id:
          type: integer
        mes:
          type: string
        ano:
          type: integer
        requisicao:
          type: ["string", "null"]
        posto:
          type: ["string", "null"]
        data:
          type: string
          format: date
        placa:
          type: string
        litros:
          type: number
        precoLitro:
          type: number
        totalPago:
          type: number
        kmInicio:
          type: ["number", "null"]
        kmFinal:
          type: ["number", "null"]
        kmPercorrido:
          type: ["number", "null"]
        media:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - mes
        - ano
        - data
        - placa
        - litros
        - precoLitro
        - totalPago
        - createdAt
        - updatedAt
    AbastecimentoInput:
      type: object
      properties:
        mes:
          type: string
        ano:
          type: integer
        requisicao:
          type: string
        posto:
          type: string
        data:
          type: string
          format: date
        placa:
          type: string
        litros:
          type: number
        precoLitro:
          type: number
        totalPago:
          type: number
        kmInicio:
          type: number
        kmFinal:
          type: number
        kmPercorrido:
          type: number
        media:
          type: number
      required:
        - mes
        - ano
        - data
        - placa
        - litros
        - precoLitro
        - totalPago
    AbastecimentoUpdate:
      type: object
      properties:
        mes:
          type: string
        ano:
          type: integer
        requisicao:
          type: string
        posto:
          type: string
        data:
          type: string
          format: date
        placa:
          type: string
        litros:
          type: number
        precoLitro:
          type: number
        totalPago:
          type: number
        kmInicio:
          type: number
        kmFinal:
          type: number
        kmPercorrido:
          type: number
        media:
          type: number
    AbastecimentoListResponse:
      type: object
      properties:
        abastecimentos:
          type: array
          items:
            $ref: "#/components/schemas/Abastecimento"
        total:
          type: integer
      required: [abastecimentos, total]
...
    DashboardResumo:
      type: object
      properties:
        totalFrete:
          type: number
        totalPedagio:
          type: number
        totalGeral:
          type: number
        totalViagens:
          type: integer
        totalPeso:
          type: number
        mediaPorViagem:
          type: number
        melhorFrota:
          type: ["string", "null"]
        totalDiesel:
          type: number
        totalLitros:
          type: number
      required:
        - totalFrete
        - totalPedagio
        - totalGeral
        - totalViagens
        - totalPeso
        - mediaPorViagem
        - totalDiesel
        - totalLitros
    FrotaRevenue:
      type: object
      properties:
        frota:
          type: string
        totalFrete:
          type: number
        totalPedagio:
          type: number
        totalGeral:
          type: number
        totalViagens:
          type: integer
        rank:
          type: integer
      required: [frota, totalFrete, totalPedagio, totalGeral, totalViagens, rank]
    PeriodoRevenue:
      type: object
      properties:
        periodo:
          type: string
        totalFrete:
          type: number
        totalPedagio:
          type: number
        totalGeral:
          type: number
        totalViagens:
          type: integer
      required: [periodo, totalFrete, totalPedagio, totalGeral, totalViagens]
    DieselByPlaca:
      type: object
      properties:
        placa:
          type: string
        totalLitros:
          type: number
        totalPago:
          type: number
        mediaGeral:
          type: ["number", "null"]
        kmTotal:
          type: ["number", "null"]
        totalAbastecimentos:
          type: integer
      required: [placa, totalLitros, totalPago, totalAbastecimentos]
    MensalComparativo:
      type: object
      properties:
        mes:
          type: string
        frete:
          type: number
        pedagio:
          type: number
        diesel:
          type: number
        lucroLiquido:
          type: number
        viagens:
          type: integer
      required: [mes, frete, pedagio, diesel, lucroLiquido, viagens]