# lib/api-zod/src/generated/types/abastecimentoInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Revenue Management API
 * OpenAPI spec version: 0.2.0
 */

export interface AbastecimentoInput {
  mes: string;
  ano: number;
  requisicao?: string;
  posto?: string;
  data: Date;
  placa: string;
  litros: number;
  precoLitro: number;
  totalPago: number;
  kmInicio?: number;
  kmFinal?: number;
  kmPercorrido?: number;
  media?: number;
}