# lib/api-zod/src/generated/types/abastecimentoListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Revenue Management API
 * OpenAPI spec version: 0.2.0
 */
import type { Abastecimento } from './abastecimento';

export interface AbastecimentoListResponse {
  abastecimentos: Abastecimento[];
  total: number;
}