# lib/api-zod/src/generated/types/dashboardResumo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Revenue Management API
 * OpenAPI spec version: 0.2.0
 */

export interface DashboardResumo {
  totalFrete: number;
  totalPedagio: number;
  totalGeral: number;
  totalViagens: number;
  totalPeso: number;
  mediaPorViagem: number;
  /** @nullable */
  melhorFrota?: string | null;
  totalDiesel: number;
  totalLitros: number;
}