# lib/api-zod/src/generated/types/dieselByPlaca.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Revenue Management API
 * OpenAPI spec version: 0.2.0
 */

export interface DieselByPlaca {
  placa: string;
  totalLitros: number;
  totalPago: number;
  /** @nullable */
  mediaGeral?: number | null;
  /** @nullable */
  kmTotal?: number | null;
  totalAbastecimentos: number;
}