# lib/api-zod/src/generated/types/frotaRevenue.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Revenue Management API
 * OpenAPI spec version: 0.2.0
 */

export interface FrotaRevenue {
  frota: string;
  totalFrete: number;
  totalPedagio: number;
  totalGeral: number;
  totalViagens: number;
  rank: number;
}