# lib/api-zod/src/generated/types/mensalComparativo.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Fleet Revenue Management API
 * OpenAPI spec version: 0.2.0
 */

export interface MensalComparativo {
  mes: string;
  frete: number;
  pedagio: number;
  diesel: number;
  lucroLiquido: number;
  viagens: number;
}